requirements.txt
flask==2.0.1
werkzeug==2.0.3
openai==0.27.8
SpeechRecognition==3.10.1
pyttsx3==2.90
python-dotenv==1.0.1
sounddevice==0.4.6
numpy==1.26.4
scipy==1.12.0
gTTS==2.3.1
gunicorn==20.1.0 
flask-cors==3.0.10